pyproject.toml
[project]
name = "revisao-sistematica-com-agentes"
version = "0.1.0"
description = "AI-powered screening agent for systematic review triage"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph==1.0.8",
    "langchain==1.2.7",
    "langgraph-prebuilt==1.0.7",
    "langchain-google-genai>=2.1",
    "langchain-ollama==1.0.1",
    "langchain-openai>=0.1",
    "xlrd>=2.0",
    "openpyxl>=3.1",
    "python-dotenv>=1.1",
    "pydantic>=2.0",
    "pandas>=2.0",
]

[project.scripts]
screening = "screening.main:main"
